Uplift Sample Upload script to binary artifact repositories
