+++ Pipeline/PackageBuildOutputs/appArtifactRepository.properties
@@ -1,0 +1,22 @@
+########################################################################
+### The following properties are required for publishing a package 
+### containing build outputs to an Artifact Repository Server 
+### such as JFrog Artifactory or Sonartype Nexus  
+########################################################################
+# The properties file is optional, while all parameters can also be specified
+# via the cli. Especially the credentials should be stored in the 
+# secret store of your pipeline orchestrator.
+########################################################################
+
+# The URL to the Artifact server
+artifactRepository.url=https://your-artifactory-url/artifactory
+
+# The artifact repository name to store the build 
+artifactRepository.repo=sys-zos-application-local
+
+# Artifact repository directory
+artifactRepository.directory=
+
+# Artifact repository credentials
+artifactRepository.user=
+artifactRepository.password=
