--- Pipeline/PackageBuildOutputs/appArtifactory.properties
@@ -1,14 +1,0 @@
-########################################################################
-### The following properties are required for publishing a package 
-### containing build outputs to an Artifactory Server.  
-########################################################################
-
-# The URL to the Artifactory Server
-artifactory.url=https://your-artifactory-url/artifactory
-
-# The Artifactory repository to store the build 
-artifactory.repo=sys-zos-application-local
-
-# Artifactory credentials
-artifactory.user=
-artifactory.password=
